pyproject.toml
[project]
name = "impactable"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "django>=4.2,<4.3",
    "django-cors-headers==3.13.0",
    "django-filter==2.2.0",
    "djangorestframework==3.14.0",
    "drf-yasg==1.21.4",
    "fastparquet>=2024.11.0,<2024.12",
    "haikunator>=2.1.0,<2.2",
    "jupyter==1.0.0",
    "matplotlib>=3.9.2,<3.10",
    "numpy>=2.2,<2.3",
    "ollama>=0.5,<0.6",
    "openpyxl==3.1.2",
    "pandas>=2.2,<2.3",
    "pantab>=5.2.0,<5.3",
    "pip-audit>=2.4,<2.5",
    "psycopg[binary]>=3.2.9,<3.3",
    "pyarrow>=19.0.0,<19.1",
    "pylint==2.17.5",
    "pytest>=9.0.2",
    "python-dateutil==2.8.2",
    "python-dotenv==0.10.3",
    "rapidfuzz>=3.13,<3.14",
    "scikit-learn>=1.5.2,<1.6",
    "scipy>=1.14.1,<1.15",
    "seaborn==0.13.0",
    "statsmodels>=0.14,<0.15",
    "wbdata>=1.0,<1.1",
    "wordcloud>=1.9.3,<1.10",
    "xlsxwriter>=3.1,<3.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
